Refactor to make the code more organize with modules
